// apps/api/src/discount/discountModel.ts
import { ObjectId } from "mongodb";

export type AdjustmentMethod = 'absolute' | 'percentage';

type Entity = 'seller' | 'user' | 'prixer' | 'organization' | 'manufacturer';

export interface ApplicableEntity {
  id?: string; // the Id of the Entity if it is a specific one instead of a group
  name?: string;
  type: Entity; // the type of entity
  adjustmentMethod: AdjustmentMethod; // absolute or percentage
  customValue?: number; // amount
}

export interface Discount {
  _id?: ObjectId;
  name: string;
  description: string;
  active: boolean;
  adjustmentMethod: AdjustmentMethod; // Defines whether the adjustment is an absolute amount or a percentage.
  defaultValue: number;  // The value to adjust by – for percentage decimals as needed.
  applicableProducts?: [string, string?][]; // Product Id + Variant Id
  applicableArts?: string[]; // applicable art IDs for surcharge
  entityOverrides?: ApplicableEntity[]; // overrides for specific users or groups of users
  appliesToAllProducts: boolean;
  appliestoAllArts: boolean;
  dateRange?: {
    start: Date;
    end: Date;
  };
}